# Conflicts:
#	src/main/java/org/hbrs/se2/project/startupx/repository/UserRepository.java

+++ src/main/java/org/hbrs/se2/project/startupx/repository/KommentarRepository.java
@@ -3,5 +3,5 @@ package org.hbrs.se2.project.startupx.repository;
 import org.hbrs.se2.project.startupx.entities.Kommentar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KommentarRepository extends JpaRepository<Kommentar, Long> {
+public interface KommentarRepository extends JpaRepository<Kommentar, Integer> {
 }
